mcp-tools/pyproject.toml
[project]
name = "mcp-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.1",
    "openai>=1.86.0",
    "pyyaml>=6.0.2",
]
